[개발]이벤트를 담당하는 객체 추가 CardEventController #11.
